-m files to /src, webpack enabled
