+++ app/views/simulator/edit.html.erb
@@ -522,15 +522,9 @@
 </script>
 
 <!-- hotkeys -->
-<script src="/js/hotkey_binder/defaultKeys.json"></script>
-<script src="/js/hotkey_binder/model/normalize/normalizer.plugin.js"></script>
-<script src="/js/hotkey_binder/model/normalize/test.js"></script>
-<script src="/js/hotkey_binder/model/utils.js"></script>
-<script src="/js/hotkey_binder/model/shortcuts.plugin.js"></script>
-<script src="/js/hotkey_binder/view/panel.ui.js"></script>
-<script src="/js/hotkey_binder/model/addShortcut.js"></script>
-<script src="/js/hotkey_binder/model/actions.js"></script>
-<script src="/js/hotkey_binder/keyBinder.js"></script>
+<script src="/src/hotkey_binder/model/normalize/normalizer.plugin.js"></script>
+<script src="/src/hotkey_binder/model/utils.js"></script>
+<script src="/src/hotkey_binder/model/shortcuts.plugin.js"></script>
 <!-- hotkeys-end --> 
 
 <script>
