+++ public/src/hotkey_binder/documentation.txt
