+++ public/src/hotkey_binder/model/Shortcuts.plugin.js
