+++ public/src/hotkey_binder/model/normalize/normalizer.plugin.js
