+++ public/src/hotkey_binder/model/utils.js
